requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
google-generativeai>=0.7.0
pydantic>=2.7.0
python-dotenv>=1.0.0
streamlit>=1.35.0
requests>=2.31.0
